debian: Build libscanmem1 and libscanmem-dev
The Debian policy requires that the libscanmem1 and libscanmem-dev
packages are built. Furthermore we've been asked to set strict
version depends to libscanmem1, to use libscanmem.so.1, to deploy
the libscanmem.so symlink with the -dev package and to install the
lib to /usr/lib/scanmem. So do it exactly like that and make lintian
happy by compiling the gameconqueror package with 'any' arch instead
of 'all'.